file_write()

+++ semestralka/include/filesystem.hpp
@@ -148,7 +148,7 @@ public:
   // offset is in bytes - offset 15 means, that 15th byte will be the first to
   // be written to
   void file_write(int32_t inode_id, int32_t offset, const char *data,
-                  int32_t data_size);
+                  size_t data_size);
   // handle read accross multiple clusters
   std::vector<uint8_t> file_read(int32_t inode_id);
   // remove file & remove from parent directory
